Find Minimum Number of Steps Required to Make all Elements Equal in Array
A Java program that finds the minimum number of steps required to make all elements equal in an array. This is done by firstly sorting the array into ascending order and by getting the middle element. The middle element is used as the other elements are the closest to it. We can loop through a for loop to count the number of steps it takes to make the rest of the elements equal.